Type Alias libbgs::markoff::Coord

pub type Coord<const P: u128> = FpNum<P>;

A coordinate for a Markoff triple. May represent any of $a$, $b$, or $c$ in a Markoff triple $(a, b, c)$. This is a single field struct containing only an FpNum<P> for prime P.

Aliased Type

struct Coord<const P: u128>(pub u128);

Fields
0: u128
Implementations

impl<const P: u128> Coord<P>

pub fn to_chi(&self) -> Either<QuadNum<P>, FpNum<P>>

Returns an element $\chi$ such that, for a coordinate $a$, $a = \chi + \chi^{-1}$. If $a$ is a quadratic residue modulo P, then $\chi \in \mathbb{F}_p$, and the result will be a Right<FpNum<P>>. Otherwise, $\chi \in \mathbb{F}_{p^2}$, and the result will be a Left<QuadNum<P>>.

pub fn from_chi_fp<S, const L: usize>( chi: &SylowElem<S, L, FpNum<P>>, decomp: &SylowDecomp<S, L, FpNum<P>> ) -> Coord<P>where FpNum<P>: Factor<S, L>,

Returns the coordinate $a = \chi + \chi^{-1}$, where $\chi \in \mathbb{F}_p$.

pub fn from_chi_quad<S, const L: usize>( chi: &SylowElem<S, L, QuadNum<P>>, decomp: &SylowDecomp<S, L, QuadNum<P>> ) -> Coord<P>where QuadNum<P>: Factor<S, L>,

Returns the coordinate $a = \chi + \chi^{-1}$, where $\chi \in \mathbb{F}_{p^2}$.

pub fn rot( self, b: Coord<P>, c: Coord<P> ) -> impl Iterator<Item = (Coord<P>, Coord<P>)>

Returns an iterator yielding the coordinates $(b, c)$ contained in the orbit with fixed coordinate $a$ (the coordinate on which rot is called), beginning with $(a, b, c)$.

pub fn get_ord<S1, S2, const L1: usize, const L2: usize>(&self) -> u128where FpNum<P>: Factor<S1, L1>, QuadNum<P>: Factor<S2, L2>,

Returns the order of the map $\text{rot}_a$, that is, $\lvert \langle \text{rot}_a \rangle \rvert$.

impl<const P: u128> FpNum<P>

pub fn int_sqrt(&self) -> Option<FpNum<P>>

Calculates this number’s square root, if it is a quadratic residue; otherwise, returns None.

pub const fn find_nonresidue() -> u128

Returns a quadratic nonresidue modulo p.