Module libbgs::numbers

Types and utilities for manipulating numbers in various types of finite fields.

This module contains modular integers (i.e., $\mathbb{Z} / p\mathbb{Z}$ for prime $p$), their quadratic finite field extensions (i.e., $\mathbb{Z} / p^2\mathbb{Z}$ for prime $p$), and decompositions into direct sums of Sylow subgroups.

Modules

flags

Bitwise flags for configuring a SylowStreamBuilder. May be combined with the bitwise integer OR operator, |.

tests

Utility methods for use in other tests. These methods should probably not be used outside of this crate.

Structs

FactorStream

An iterator yielding all of the factors of some number beneath a limit. The type parameter L is the length of the factorization.

Factorization

A prime power decomposition of a positive integer.

FpNum

An integer modulo P.

QuadField

The finite field of size P^2. Isomorphic to $\mathbb{Z} / P^2\mathbb{Z}$. Each QuadField has a fixed quadratic nonresidue r used as a basis element for the numbers outside of the prime subfield.

QuadNum

An integer modulo P^2. An element $x$ is represented as $x = a_0 + a_1\sqrt{r}$, where $r$ is the fixed basis element.

SylowDecomp

A decomposition of a finite cyclic group into the direct sum of its Sylow subgroups. In particular, this group represents the right hand side of the isomorphism $$G \cong \bigoplus_{i = 1}^n \mathbb{Z} / p_i^{t_i} \mathbb{Z}$$ where $$|G| = \prod_{i = 1}^n p_i^{t_i}$$ and $G$ is a finite cyclic group.

SylowElem

An element of the decomposition of a finite cyclic group into the direct sum of its Sylow subgroups.

SylowParStream

A stream yielding elements of particular orders, as their Sylow decompositions. Generates the elements in parallel on multiple threads.

SylowSeqStream

A stream yielding elements of particular orders, as their Sylow decompositions. Generates the elements sequentially on a single thread.

SylowStreamBuilder

A builder for a stream yielding elements of particular orders, as their Sylow decompositions.

Traits

Factor

Types that have a size or order which can be expressed as a product of prime powers. The type parameter S is a phantom type to allow users of this library to provide their own factorizations for FpNum<P>, QuadNum<P>, etc. for arbitrary P. The type S must come from the crate implementing the factorization to satisfy Rust’s coherence and orphan rules, but does not affect the memory layout of any instance of an object implementing Factor. See the Rust documentation on coherence and the orphan rule.

GroupElem

Types that represent the elements of a group. In order for a type to represent the elements of the group, the type must satisfy these axioms:

SylowDecomposable

Groups that can be decomposed into a direct sum of cyclic Sylow subgroups. In particular, these groups must be finite and cyclic.