Module libbgs::numbers::flags

Bitwise flags for configuring a SylowStreamBuilder. May be combined with the bitwise integer OR operator, |.

Constants

INCLUDE_ONE

Stream is guaranteed to include the identity.

LEQ

Stream will yield any element less than or equal to a target. Guaranteed to only return 1 value of \chi for each threat; guaranteed 1 for 1.

NONE

Stream will behave with all default options. Identity when used with the | operator.

NO_PARABOLIC

Stream will yield elements in a parabolic order.

NO_UPPER_HALF

Stream will yield half as many elements as the default. Precisely, the stream will yield either $\chi$ or $\chi^{-1}$, but not both, for every $\chi \in F_{p^2}$. Even more precisely, for every $x$ the builder yields, the first non-zero coordinate is guarantied to be less than half the maximum of the its corresponding prime power.