Struct libbgs::numbers::QuadNum

pub struct QuadNum<const P: u128>(pub FpNum<P>, pub FpNum<P>);

An integer modulo P^2. An element $x$ is represented as $x = a_0 + a_1\sqrt{r}$, where $r$ is the fixed basis element.

The association between a QuadNum instance and a QuadField (hence, the fixed basis element r), is implicit. You must ensure that operations on QuadNum instances are associated to the same choice of r.

Tuple Fields
0: FpNum<P>

The value $a_0$, when writing this QuadNum as $a_0 + a_1\sqrt{r}$.

1: FpNum<P>

The value $a_1$, when writing this QuadNum as $a_0 + a_1\sqrt{r}$.

Implementations

impl<const P: u128> QuadNum<P>

pub fn is_zero(&self) -> bool

True if this number is zero; false otherwise.

pub fn steinitz(i: u128) -> QuadNum<P>

Returns the Steinitz element of $\mathbb{F}_{p^2}$ with index i.

pub fn int_sqrt_either(x: FpNum<P>) -> Either<QuadNum<P>, FpNum<P>>

Calculates the square root of an integer modulo P, casting to an FpNum<P> if x is a quadratic residue. Returns a Left QuadNum<P> if x is a quadratic nonresidue, or a Right FpNum<P> if x is a quadratic residue (including 0).

pub fn int_sqrt(x: FpNum<P>) -> QuadNum<P>

Calculates the square root af in integer modulo P.

pub const R: FpNum<P> = _

The basis element for the numbers outside of the prime subfield.