Trait libbgs::numbers::GroupElem

pub trait GroupElem: Clone + PartialEq + Eq + Debug {
    // Required methods
    fn size() -> u128;
    fn one() -> Self;
    fn is_one(&self) -> bool;
    fn multiply(&self, other: &Self) -> Self;

    // Provided methods
    fn pow(&self, n: u128) -> Self { ... }
    fn inverse(&self) -> Self { ... }
    fn order<S, const L: usize>(&self) -> Factorization<L>
       where Self: Factor<S, L> { ... }
    fn count_elements_of_order<S, const L: usize>(ds: &[u128; L]) -> u128
       where Self: SylowDecomposable<S, L> { ... }
}

Types that represent the elements of a group. In order for a type to represent the elements of the group, the type must satisfy these axioms:

The type has a binary operator (multiply).

The type must have a unique element given by one().

Event element of the type must have element given by inverse().

Required Methods

fn size() -> u128

Gets the size of the group this element belongs to.

fn one() -> Self

Gets the unique identity element of this group.

fn is_one(&self) -> bool

True if this element is the multiplicative identity; false otherwise.

fn multiply(&self, other: &Self) -> Self

Returns the product of two elements under the group binary operator. If you implement this trait, you must guarantee that the operation is associative; that is, a.multiply(b.multiply(c, &g), &g) == a.multiply(b, &g).multiply(c, &g).

Provided Methods

fn pow(&self, n: u128) -> Self

Raises this element to the power of n. If you override this trait, you must guarantee that x.pow(2) == x.square() for all x.

fn inverse(&self) -> Self

Returns the multiplicative inverse of this element. If you implement this trait, you must guarantee x.inverse().multiply(x) and x.multiply(x.inverse()) both evaluate to the unique identity element.

fn order<S, const L: usize>(&self) -> Factorization<L>where Self: Factor<S, L>,

Returns the order of this element, that is, the smallest positive power p for which a.pow(p, &g).is_one(&g) returns True.

fn count_elements_of_order<S, const L: usize>(ds: &[u128; L]) -> u128where Self: SylowDecomposable<S, L>,

Returns the number of elements of a particular order. The argument is the powers of the prime factors of the group’s order.